Function kanal::unbounded

pub fn unbounded<T>() -> (Sender<T>, Receiver<T>)

Creates a new sync unbounded channel, and returns Sender and Receiver of the channel for type T, you can get access to async API of AsyncSender and AsyncReceiver with to_sync, as_async or clone_sync based on your requirements, by calling them on sender or receiver.

Examples

use std::thread::spawn;

let (s, r) = kanal::unbounded(); // for channel with unbounded size queue, this channel never blocks on send

// spawn 8 threads, that will send 100 numbers to the channel reader
for i in 0..8{
    let s = s.clone();
    spawn(move || {
        for i in 1..100{
            s.send(i);
        }
    });
}
// drop local sender so the channel send side gets closed when all of the senders finished their jobs
drop(s);

let first = r.recv().unwrap(); // receive first msg
let total: u32 = first+r.sum::<u32>(); // the receiver implements iterator so you can call sum to receive sum of rest of messages
assert_eq!(total, 39600);