Struct kanal::AsyncReceiver

pub struct AsyncReceiver<T> { /* private fields */ }

AsyncReceiver is receiving side of the channel in async mode. Receivers can be cloned and produce receivers to operate in both sync and async modes.

Examples

let (_s, receiver) = kanal::bounded_async::<u64>(0);
let sync_receiver=receiver.clone_sync();

Implementations

impl<T> AsyncReceiver<T>

pub fn recv(&self) -> ReceiveFuture<'_, T>

Returns a ReceiveFuture to receive data from the channel asynchronously.

Examples

let name=r.recv().await?;
println!("Hello {}",name);

pub fn stream(&self) -> ReceiveStream<'_, T>

Creates a asynchronous stream for the channel to receive messages, ReceiveStream borrows the AsyncReceiver, after dropping it, receiver will be available and usable again.

Examples

// import to be able to use stream.next() function
use futures::stream::StreamExt;
// import to be able to use stream.is_terminated() function
use futures::stream::FusedStream;

let (s, r) = kanal::unbounded_async();
co(async move {
    for i in 0..100 {
        s.send(i).await.unwrap();
    }
});
let mut stream = r.stream();
assert!(!stream.is_terminated());
for i in 0..100 {
    assert_eq!(stream.next().await, Some(i));
}
// Stream will return None after it is terminated, and there is no other sender.
assert_eq!(stream.next().await, None);
assert!(stream.is_terminated());

pub fn try_recv(&self) -> Result<Option<T>, ReceiveError>

Tries receiving from the channel without waiting on the waitlist. It returns Ok(Some(T)) in case of successful operation and Ok(None) for a failed one, or error in case that channel is closed. Important note: this function is not lock-free as it acquires a mutex guard of the channel internal for a short time.

Examples

loop {
    if let Some(name)=r.try_recv()?{
        println!("Hello {}!",name);
        break;
    }
}

pub fn try_recv_realtime(&self) -> Result<Option<T>, ReceiveError>

Tries receiving from the channel without waiting on the waitlist or waiting for channel internal lock. It returns Ok(Some(T)) in case of successful operation and Ok(None) for a failed one, or error in case that channel is closed. Do not use this function unless you know exactly what you are doing.

Examples

loop {
    if let Some(name)=r.try_recv_realtime()?{
        println!("Hello {}!",name);
        break;
    }
}

pub fn is_disconnected(&self) -> bool

Returns, whether the send side of the channel, is closed or not.

Examples

let (s, r) = kanal::unbounded::<u64>();
drop(s); // drop sender and disconnect the send side from the channel
assert_eq!(r.is_disconnected(),true);

pub fn is_terminated(&self) -> bool

Returns, whether the channel receive side is terminated, and will not return any result in future recv calls.

Examples

let (s, r) = kanal::unbounded::<u64>();
s.send(1).unwrap();
drop(s); // drop sender and disconnect the send side from the channel
assert_eq!(r.is_disconnected(),true);
// Also channel is closed from send side, it's not terminated as there is data in channel queue
assert_eq!(r.is_terminated(),false);
assert_eq!(r.recv().unwrap(),1);
// Now channel receive side is terminated as there is no sender for channel and queue is empty
assert_eq!(r.is_terminated(),true);

pub fn clone_sync(&self) -> Receiver<T>

Returns sync cloned version of the receiver.

Examples

let (s, r) = kanal::unbounded_async();
s.send(1).await?;
let sync_receiver=r.clone_sync();
// JUST FOR EXAMPLE IT IS WRONG TO USE SYNC INSTANCE IN ASYNC CONTEXT
assert_eq!(sync_receiver.recv()?,1);

pub fn to_sync(self) -> Receiver<T>

Converts AsyncReceiver to Receiver and returns it.

Examples

  let (s, r) = kanal::bounded_async(0);
  // move to sync environment
  std::thread::spawn(move || {
    let r=r.to_sync();
    let name=r.recv()?;
    println!("Hello {}!",name);
    anyhow::Ok(())
  });
  s.send("World").await?;

pub fn as_sync(&self) -> &Receiver<T>

Borrows AsyncReceiver as Receiver and returns it

Examples

  let (s, r) = kanal::bounded_async(0);
  // move to sync environment
  std::thread::spawn(move || {
    let name=r.as_sync().recv()?;
    println!("Hello {}!",name);
    anyhow::Ok(())
  });
  s.send("World").await?;

pub fn is_bounded(&self) -> bool

Returns whether the channel is bounded or not.

Examples

let (s, r) = kanal::bounded::<u64>(0);
assert_eq!(s.is_bounded(),true);
assert_eq!(r.is_bounded(),true);

let (s, r) = kanal::unbounded::<u64>();
assert_eq!(s.is_bounded(),false);
assert_eq!(r.is_bounded(),false);

pub fn len(&self) -> usize

Returns length of the queue.

Examples

let (s, r) = kanal::unbounded::<u64>();
assert_eq!(s.len(),0);
assert_eq!(r.len(),0);
s.send(10);
assert_eq!(s.len(),1);
assert_eq!(r.len(),1);

pub fn is_empty(&self) -> bool

Returns whether the channel queue is empty or not.

Examples

let (s, r) = kanal::unbounded::<u64>();
assert_eq!(s.is_empty(),true);
assert_eq!(r.is_empty(),true);

pub fn is_full(&self) -> bool

Returns whether the channel queue is full or not full channels will block on send and recv calls it always returns true for zero sized channels.

Examples

let (s, r) = kanal::bounded(1);
s.send("Hi!").unwrap();
assert_eq!(s.is_full(),true);
assert_eq!(r.is_full(),true);

pub fn capacity(&self) -> usize

Returns capacity of channel (not the queue) for unbounded channels, it will return usize::MAX.

Examples

let (s, r) = kanal::bounded::<u64>(0);
assert_eq!(s.capacity(),0);
assert_eq!(r.capacity(),0);

let (s, r) = kanal::unbounded::<u64>();
assert_eq!(s.capacity(),usize::MAX);
assert_eq!(r.capacity(),usize::MAX);

pub fn receiver_count(&self) -> u32

Returns count of alive receiver instances of the channel.

Examples

let (s, r) = kanal::unbounded::<u64>();
let receiver_clone=r.clone();
assert_eq!(r.receiver_count(),2);

pub fn sender_count(&self) -> u32

Returns count of alive sender instances of the channel.

Examples

let (s, r) = kanal::unbounded::<u64>();
let sender_clone=s.clone();
assert_eq!(r.sender_count(),2);

pub fn close(&self) -> bool

Closes the channel completely on both sides and terminates waiting signals.

Examples

let (s, r) = kanal::unbounded::<u64>();
// closes channel on both sides and has same effect as r.close();
s.close();
assert_eq!(r.is_closed(),true);
assert_eq!(s.is_closed(),true);

pub fn is_closed(&self) -> bool

Returns whether the channel is closed on both side of send and receive or not.

Examples

let (s, r) = kanal::unbounded::<u64>();
// closes channel on both sides and has same effect as r.close();
s.close();
assert_eq!(r.is_closed(),true);
assert_eq!(s.is_closed(),true);